[ADD] Only the map as argument = no menu. One more random argument = menu.

+++ src/death.c
@@ -76,7 +76,7 @@ void		death(t_env *env)
 
 	i = 0;
 	if (env->player.nb_shots > 0)
-		env->player.accuracy = (int)((env->player.touched / env->player.nb_shots) * 100);
+		env->player.accuracy = ((env->player.touched / env->player.nb_shots) * 100);
 	if (!env->confirmation_box.state)
 	{
 		SDL_SetRelativeMouseMode(0);
